requirements.txt
camelot-py==0.9.0
numpy==1.25.2
pandas==2.1.0
